requirements.txt
# Core dependencies
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LLM and transformers
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Web API
flask>=2.3.0
requests>=2.31.0

# Testing
pytest>=7.0.0
